openseneca/requirements.txt
fire>=0.6.0
uvicorn[standard]>=0.29.0
fastapi>=0.110.1
pytest
pandas>=2.2.1
torch==2.2.1
transformers>=4.38.2
safetensors
scikit-learn>=1.4.0
evaluate>=0.4.1
lm_eval>=0.4.2
langchain>=0.1.13
openai>=1.14.2
langchain_openai>=0.1.1
tqdm
pylint
pytest-cov
colorlog
pycountry
python-dotenv
langdetect
pydantic
huggingface-cli